% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_course_sheet.R
\name{get_r_course_gsheet}
\alias{get_r_course_gsheet}
\title{Get R course sheet}
\usage{
get_r_course_gsheet()
}
\description{
Get R course sheet
}
\details{
Get the R course sheet from Google sheets.
}
