% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stima_r_courses.R
\name{get_stima_r_courses}
\alias{get_stima_r_courses}
\title{Get the courses that have R package dependencies}
\usage{
get_stima_r_courses()
}
\description{
Get the courses that have R package dependencies
}
