% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_stima_packages.R
\name{install_stima_packages}
\alias{install_stima_packages}
\title{Install Stima packages}
\usage{
install_stima_packages(course = NULL)
}
\arguments{
\item{course}{What course to install for. See \code{get_stima_r_courses()}.}
}
\description{
Install Stima packages
}
\details{
The package is based on the Stima R Package google sheets
}
\seealso{
get_stima_r_courses
}
